+++ libraries/classes/Plugins/Export/Helpers/TableProperty.php
@@ -83,7 +83,7 @@ class TableProperty
      */
     public function getPureType()
     {
-        $pos = mb_strpos($this->type, '(');
+        $pos = (int) mb_strpos($this->type, '(');
         if ($pos > 0) {
             return mb_substr($this->type, 0, $pos);
         }
